feat: split dataset into train and test

+++ model/train.py
@@ -14,7 +14,8 @@ from keras.models import Model, Sequential
 from keras.utils import to_categorical
 from keras.layers import Embedding, concatenate, LSTM, Dropout, Input, Reshape, Dense
 from keras.callbacks import ModelCheckpoint, TensorBoard
-from utils import get_max_seq_length, encode_and_pad, pad_left
+from utils import pad_left, prepare_dataset
+from sklearn.model_selection import train_test_split
 
 pp = PrettyPrinter(indent=2)
 
@@ -23,13 +24,21 @@ def pprint(obj):
 
 # --- Command Line Arguments ---
 
+boolean = lambda x: (str(x).lower() == 'true')
+
 parser = ArgumentParser()
 parser.add_argument("--max-seq-len", nargs='?', type=int, const=True, default=100)
 parser.add_argument("--epochs", nargs='?', type=int, const=True, default=300)
 parser.add_argument("--batch-size", nargs='?', type=int, const=True, default=64)
 parser.add_argument("--name", nargs='?', type=str, const=True)
+parser.add_argument("--dry-run", nargs='?', type=boolean, const=True, default=False)
 
 args = parser.parse_args()
+
+dry_run = args.dry_run
+
+if dry_run:
+  print("[INFO] Running script in dry mode.")
 
 # --- Parameters ---
 
@@ -39,19 +48,20 @@ n_epochs = args.epochs
 
 # --- Create Directory ---
 
-out_dir = "./runs"
+if not dry_run:
+  out_dir = "./runs"
 
-if not path.exists(out_dir):
-  makedirs(out_dir)
+  if not path.exists(out_dir):
+    makedirs(out_dir)
 
-n_runs = len(listdir(out_dir)) + 1
-run_name = args.name if args.name else "run_" + pad_left(n_runs)
-run_dir = out_dir + "/" + run_name
+  n_runs = len(listdir(out_dir)) + 1
+  run_name = args.name if args.name else "run_" + pad_left(n_runs, 3)
+  run_dir = out_dir + "/" + run_name
 
-if not path.exists(run_dir):
-  makedirs(run_dir)
-else:
-  raise Exception("Cannot create log directory for run. Directory already exists.")
+  if not path.exists(run_dir):
+    makedirs(run_dir)
+  else:
+    raise Exception("Cannot create log directory for run. Directory already exists.")
 
 # --- Data Preprocessing ---
 
@@ -83,34 +93,14 @@ sequences = tokenizer.texts_to_sequences(inputs)
 function_signatures = sequences[:n_observations]
 function_bodies = sequences[n_observations:]
 
-# signature length is fixed
-max_signature_seq_length = get_max_seq_length(function_signatures, max_seq_len)
-max_body_seq_length = get_max_seq_length(function_bodies, max_seq_len)
+# train, test split
+seed = 1
+test_size=0.33
+x1_train, x1_test, x2_train, x2_test = train_test_split(function_signatures, function_bodies, test_size=test_size, random_state=seed)
 
 # finalize inputs to the model
-X1_signatures, X2_bodies, y = list(), list(), list()
-for body_idx, seq in enumerate(function_bodies):
-
-  encoded_signature = encode_and_pad(function_signatures[body_idx], max_signature_seq_length, vocab_size)
-
-  for i in range(1, len(seq)):
-    # add function signature
-    X1_signatures.append(encoded_signature)
-
-    # add the entire sequence to the input and only keep the next word for the output
-    in_seq, out_seq = seq[:i], seq[i]
-
-    # apply padding and encode sequence
-    in_seq = pad_sequences([in_seq], maxlen=max_body_seq_length)[0]
-
-    # one hot encode output sequence
-    out_seq = to_categorical([out_seq], num_classes=vocab_size)[0]
-    y.append(out_seq)
-
-    # cut the input seq to fixed length
-    X2_bodies.append(in_seq[-max_seq_len:])
-
-X1_signatures, X2_bodies, y = np.array(X1_signatures), np.array(X2_bodies), np.array(y)
+x1_train, x2_train, y_train = prepare_dataset(x1_train, x2_train, max_seq_len, vocab_size)
+x1_test, x2_test, y_test = prepare_dataset(x1_test, x2_test, max_seq_len, vocab_size)
 
 # --- Information ---
 
@@ -123,10 +113,12 @@ infos.align["Value"] = "l"
 infos.add_row(["Epochs", n_epochs])
 infos.add_row(["Batch Size", batch_size])
 infos.add_row(["Max Sequence", max_seq_len])
-infos.add_row(["Observations", max(len(function_signatures), len(function_bodies))])
-infos.add_row(["X1 (Signatures)", X1_signatures.shape])
-infos.add_row(["X2 (Bodies)", X2_bodies.shape])
-infos.add_row(["Y", y.shape])
+infos.add_row(["Test Size", test_size])
+infos.add_row(["Observations (Train)", max(len(x1_train), len(x2_train))])
+infos.add_row(["Observations (Test)", max(len(x1_test), len(x2_test))])
+infos.add_row(["X1 (Signatures)", x1_train.shape])
+infos.add_row(["X2 (Bodies)", x2_train.shape])
+infos.add_row(["Y", y_train.shape])
 infos.add_row(["Vocabulary", vocab_size])
 
 print(infos)
@@ -134,11 +126,11 @@ print(infos)
 # --- Model ---
 
 # encoder
-x1_input = Input(shape=X1_signatures[0].shape, name="x1_input")
+x1_input = Input(shape=x1_train[0].shape, name="x1_input")
 x1_model = LSTM(256, return_sequences=True, name="x1_lstm_1")(x1_input)
 x1_model = Dense(128, activation='relu')(x1_model)
 
-x2_input = Input(shape=X2_bodies[0].shape, name="x2_input")
+x2_input = Input(shape=x2_train[0].shape, name="x2_input")
 x2_model = Embedding(vocab_size, 200, input_length=max_seq_len)(x2_input)
 x2_model = LSTM(256, return_sequences=True, name="x2_lstm_1")(x2_model)
 x2_model = LSTM(256, return_sequences=True, name="x2_lstm_2")(x2_model)
@@ -153,19 +145,22 @@ decoder_output = Dense(vocab_size, activation='softmax')(decoder)
 model = Model(inputs=[x1_input, x2_input], outputs=decoder_output)
 model.compile(loss='categorical_crossentropy', optimizer='rmsprop', metrics=['accuracy'])
 
-# callbacks
-tensorboard = TensorBoard(run_dir)
-callbacks = [tensorboard]
+if not dry_run:
+  # set up callbacks
+  tensorboard = TensorBoard(run_dir)
+  callbacks = [tensorboard]
 
-# fit the model
-print("Buckle up and hold tight! We are about to start the training...")
-model.fit([X1_signatures, X2_bodies], y, epochs=n_epochs, batch_size=batch_size, shuffle=False, callbacks=callbacks)
+  # fit the model
+  print("Buckle up and hold tight! We are about to start the training...")
+  validation_data = ([x1_test, x2_test], y_test)
+  model.fit([x1_train, x2_train], y_train, validation_data=validation_data, epochs=n_epochs,
+            batch_size=batch_size, shuffle=False, callbacks=callbacks)
 
-# save model
-model.save(run_dir + '/model.h5')
-print("Model successfully saved.")
+  # save model
+  model.save(run_dir + '/model.h5')
+  print("Model successfully saved.")
 
-# save tokenizer
-with open('tokenizer.pickle', 'wb') as handle:
-  pickle.dump(tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)
-  print("Tokenizer successfully dumped.")
+  # save tokenizer
+  with open(run_dir + '/tokenizer.pickle', 'wb') as handle:
+    pickle.dump(tokenizer, handle, protocol=pickle.HIGHEST_PROTOCOL)
+    print("Tokenizer successfully dumped.")
